Add a "content" parameter to SubscriptionExpiration
Hardcoded for now with a "SevenDay" value
